Update of instance counting
First, if a commit sequence is not created actually, because of the
unavailability of the start commit, the instance counter, which was
increased during constructor call, will now be decreased again before
the actual creation is terminated (without any commit).
Second, a static method allows resetting the instance counter back to
"0", if needed.