Update of writing commit sequence files
Instead of constantly opening and closing a file stream for each commit
in a sequence, we now use a file channel, which is opened once at the
beginning and closed after no parent commits are available anymore.
This change also requires some changes in the test cases to handle the
additional exceptions, which may be thrown.